// frontend/src/api/models/ModelsTaskByIdResponse.ts
export interface ModelsTaskByIdResponse {
    /**
     * 
     * @type {PgtypeTimestamp}
     * @memberof ModelsTaskByIdResponse
     */
    createdAt?: PgtypeTimestamp;
    /**
     * 
     * @type {string}
     * @memberof ModelsTaskByIdResponse
     */
    description?: string;
    /**
     * 
     * @type {number}
     * @memberof ModelsTaskByIdResponse
     */
    difficulty?: number;
    /**
     * 
     * @type {number}
     * @memberof ModelsTaskByIdResponse
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof ModelsTaskByIdResponse
     */
    passes?: number;
    /**
     * 
     * @type {number}
     * @memberof ModelsTaskByIdResponse
     */
    score?: number;
    /**
     * 
     * @type {Array<DbTaskTestCase>}
     * @memberof ModelsTaskByIdResponse
     */
    testCases?: Array<DbTaskTestCase>;
    /**
     * 
     * @type {string}
     * @memberof ModelsTaskByIdResponse
     */
    title?: string;
    /**
     * 
     * @type {Array<number>}
     * @memberof ModelsTaskByIdResponse
     */
    topics?: Array<number>;
    /**
     * 
     * @type {PgtypeTimestamp}
     * @memberof ModelsTaskByIdResponse
     */
    updatedAt?: PgtypeTimestamp;
    /**
     * 
     * @type {number}
     * @memberof ModelsTaskByIdResponse
     */
    userId?: number;
}